check to make sure file exists before using os.remove

+++ ares/inference/ModelFit.py
@@ -1017,18 +1017,26 @@ class ModelFit(FitBase):
             # These suffixes are always the same
             for suffix in ['logL', 'chain', 'facc', 'pinfo', 'rinfo', 
                 'binfo', 'setup', 'load', 'fail', 'timeout']:
-                os.remove('{0!s}.{1!s}.pkl'.format(self.prefix, suffix))
+                if os.path.exists(_fn):
+                    os.remove('{0!s}.{1!s}.pkl'.format(self.prefix, suffix))
                 
                 for _fn in glob.glob('{0!s}.*.{1!s}.pkl'.format(self.prefix,\
                     suffix)):
-                    os.remove(_fn)
                     
-            os.remove('{!s}.prior_set.hdf5'.format(self.prefix))
+                    if os.path.exists(_fn):
+                        os.remove(_fn)
+            
+            if os.path.exists(_fn):        
+                os.remove('{!s}.prior_set.hdf5'.format(self.prefix))
+                
             # These suffixes have their own suffixes
             for _fn in glob.glob('{!s}.blob_*.pkl'.format(self.prefix)):
-                os.remove(_fn)
+                if os.path.exists(_fn):
+                    os.remove(_fn)
             for _fn in glob.glob('{!s}.*.blob_*.pkl'.format(self.prefix)):
-                os.remove(_fn)
+                if os.path.exists(_fn):
+                    os.remove(_fn)
+                    
         # Each processor gets its own fail file
         f = open('{!s}.fail.pkl'.format(prefix_by_proc), 'wb')
         f.close()
